solved

+++ multiplication.cpp
@@ -3,5 +3,5 @@
 #include <stdexcept>
 
 int64_t Multiply(int a, int b) {
-    return a*b;
+    return int64_t(a) * b;
 }
